--- routes/articles.js
@@ -1,10 +1,0 @@
-var express = require('express');
-var router = express.Router();
-
-/* GET home page. */
-router.get('/', function(req, res, next) {
-  res.render('index', { title: 'How-To-Nerd' });
-  
-});
-
-module.exports = router;
