+++ routes/index.js
@@ -1,10 +1,10 @@
 var express = require('express');
 var router = express.Router();
-var stormpath = require('express-stormpath');
 
 /* GET home page. */
-router.get('/', stormpath.loginRequired, function(req, res, next) {
-  res.send('Hi ' + req.user.givenName + '!');
+router.get('/', function(req, res, next) {
+  res.render('index', { title: 'How-To-Nerd' });
+  
 });
 
 module.exports = router;
