+++ views/index.hbs
@@ -3,7 +3,7 @@
 
 <div id="categories">
 <h3>Categories</h3>
-	<ul>
+	<ul style="display: flex;">
   		<li><a href="#home">Hardware</a></li>
   		<li><a href="#news">Software</a></li>
   		<li><a href="#contact">Web</a></li>
